Feat: 경로탐색 및 메뉴 추천 기능

+++ src/main/java/com/capstone/restArea/restArea/entity/RestArea.java
@@ -7,6 +7,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -35,10 +37,10 @@ public class RestArea extends BaseEntity {
     private String address;
 
     @Column(precision = 10, scale = 7)
-    private Double latitude;
+    private BigDecimal latitude;
 
     @Column(precision = 10, scale = 7)
-    private Double longitude;
+    private BigDecimal longitude;
 
     // ----- 신규 추가된 컬럼 및 관계 -----
